requirements.txt
# requirements.txt
tweepy==3.10.0
discord.py==1.7.3
web3==5.23.0
eth-account==0.5.6
eth-abi==2.1.1
telegram-python-bot==13.7
beautifulsoup4==4.10.0
requests==2.26.0
aiogram
pynacl
pyrogram

